refactor(membership): move add and edit membership plans to dashboard and add sidebar
